DateTime immutable doesn't use formatted timezone
DateTimeImmutableMock is created by formatting a new DateTimeMock, and thus avoiding code duplication.

However, sometimes the timezone can be passed in the date format (for example: "2022-03-15 14:06:30+02:00").

The previous format used to create the immutable date from the mutable one (which was "Y-m-d H:i:s.u") doesn't include the date timezone. 
Because of this, the timezone of the immutable date is lost when it is passed in a formatted date.