Handle 'U' format in createFromFormat.
Symfony 4.4 uses this strange construction to get the current time:

    DateTime::createFromFormat('U', time())

In normal PHP that works fine, but the mocked version was throwing

    TypeError: DateTime::createFromFormat() expects parameter 2 to be string, int given

This explicitly stringifies the datetime parameter to match the behavior
of the vanilla implementation. It also simplifies the workaround for
createFromFormat returning the current time when the format doesn't
contain any time parts. The prior overcomplicated version didn't work
with the relative date structure returned for the 'U' format.